AI-Training/requirements.txt
torch
torchvision
torchaudio
transformers>=4.30.0
datasets>=2.10.0
peft>=0.4.0
accelerate>=0.21.0
bitsandbytes>=0.39.0
safetensors>=0.3.0
tqdm
openai>=1.0.0
foundry-local-sdk